file structre define
