--- src/App.css
@@ -1,32 +1,0 @@
-.App {
-  text-align: center;
-  background-color: white;
-  height: 607px;
-  
-}
-
-
-@media (prefers-reduced-motion: no-preference) {
-
-
-.App-header {
-  min-height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  /* justify-content: center; */
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-}
