+++ src/components/Header/styles.scss
@@ -1,0 +1,34 @@
+@import '../../sass/mixins.scss';
+
+.forum-header {
+  display: flex;
+  flex-grow: 1;
+  a {
+    text-decoration: none;
+    color: black;
+  }
+  .header{
+      font-size: 15px;
+      display: flex;
+      justify-content: end !important;
+      justify-items: flex-end !important;
+      align-items: flex-end !important;
+      color: black;
+  }
+  @include breakpoint-mobile {
+
+  }
+  @include breakpoint-mobile-small {
+
+  }
+
+  .sub-class { // subclass of parent class
+    // normal screen size css  
+    @include breakpoint-mobile { // mobile view css 
+
+    }
+    @include breakpoint-mobile-small { // small mobile view css
+
+    }
+  }
+}
