+++ src/components/LoginForm/index.js
@@ -45,34 +45,34 @@ const Login = (props) => {
     const [showPassword, setShowPassword] = useState(false);
     const [email, setemail] = useState('')
     const [password, setpassword] = useState('')
-    
+
     const handelLogin = () => {
         const data = {
             email: email,
             password: password
         }
         axios.post('api/login', data)
-        .then(res => {
-            if (res.data.token !== undefined) {
-                localStorage.setItem("token", res.data.token);
-                axios.post('api/token/verify',{token:res.data.token})
-                .then(res => {
-                    if(res.data){
-                        localStorage.setItem('userDetails', JSON.stringify(res.data))
-                        console.log(props)
-                        props.history.push('/query')
-                    } else{
-                        alert('User not found')
-                    }
-                })
-            }
-        }).catch(err => console.log(err))
+            .then(res => {
+                if (res.data.token !== undefined) {
+                    localStorage.setItem("token", res.data.token);
+                    axios.post('api/token/verify', { token: res.data.token })
+                        .then(res => {
+                            if (res.data) {
+                                localStorage.setItem('userDetails', JSON.stringify(res.data))
+                                console.log(props)
+                                props.history.push('/query')
+                            } else {
+                                alert('User not found')
+                            }
+                        })
+                }
+            }).catch(err => console.log(err))
     }
 
     return (
         <Grid className="login-form">
             <div className="sub-class">
-
+                
             </div>
             <Paper elevation={10} style={paperStyle}>
                 <Grid align='center'>
